Restructured test directories
