+++ test/display/console_HelloWorld_tests.h
@@ -1,5 +1,5 @@
 #pragma once
 
-#include "munit.h"
+#include "../munit.h"
 
 extern MunitTest console_HelloWorld_tests[];
