+++ test/display/cursor_mock.c
