+++ test/display/cursor_mock.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "../src/shared/types.h"
+#include "../../src/shared/types.h"
 
 
 extern int cursor_mock_x;
