+++ test/display/framebuffer_tests.c
@@ -1,5 +1,5 @@
 #include "framebuffer_tests.h"
-#include "../src/display/framebuffer.h"
+#include "../../src/display/framebuffer.h"
 
 MunitResult framebuffer_should_contain_hello_test(const MunitParameter params[], void* user_data_or_fixture)
 {
